fix. 오류 파일 수정

+++ src/main/java/clofi/codeython/room/repository/RoomRepository.java
@@ -8,7 +8,7 @@ import clofi.codeython.room.domain.Room;
 import clofi.codeython.room.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     boolean existsByRoomName(String roomName);
     Room findByInviteCode(String inviteCode);
 }
